refactor: make search option code into function and move it to `cli.rs` module
